Create full_post view for rendering the full blog post

+++ blog/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import BlogPost
 from .forms import BlogForm
 
@@ -13,9 +14,13 @@ def blog_post_page(request):
     }
     return render(request, 'blog/blog.html', context)
 
-
+@login_required
 def upload_blog(request):
-    """Upload a blog to the database"""
+    """Upload a blog to the database, admin only"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Only admin can upload a blog.')
+        return redirect(reverse('blog_post_page'))
+
     if request.method == 'POST':
         form = BlogForm(request.POST, request.FILES)
         if form.is_valid():
@@ -35,9 +40,13 @@ def upload_blog(request):
 
     return render(request, template, context)
 
-
+@login_required
 def edit_blog(request, blog_id):
-    """Edit a blog"""
+    """Edit a blog, admin only"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Only admin can edit a blog.')
+        return redirect(reverse('blog_post_page'))
+
     blog = get_object_or_404(BlogPost, pk=blog_id)
     if request.method == 'POST':
         form = BlogForm(request.POST, request.FILES, instance=blog)
@@ -61,9 +70,23 @@ def edit_blog(request, blog_id):
 
     return render(request, template, context)
 
-
+@login_required
 def delete_blog(request, blog_id):
+    """Delete a blog from database, admin only"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Only admin can delete a blog.')
+        return redirect(reverse('blog_post_page'))
     blog = get_object_or_404(BlogPost, pk=blog_id)
     blog.delete()
     messages.success(request, 'Blog has been deleted')
     return redirect(reverse('blog_post_page'))
+
+
+def full_post(request, blog_id):
+    blogs = get_object_or_404(BlogPost, pk=blog_id)
+
+    context = {
+        'blogs': blogs,
+    }
+
+    return render(request, 'blog/full_blog.html', context)
